backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
